Update assignment week 2
Task 1

In this task you will construct the form for users to submit their
comments. You need to complete the following:

Set up the form with the three fields: name, rating and comment.
Use a JavaScript object in the controller to keep track of the
information. The rating should be set to 5 by default. Use radio for
specifying the rating in the form.
Use two-way data binding between the form fields and the JavaScript
object.
Task 2

In this task, you will enable form validation using Angular support for
form validation. You need to complete the following:

The name and comments are required fields and the user has entered
appropriate information in both the fields.
The submit button should be disabled until the user has entered all the
required fields.
The user should be alerted by showing the invalid fields in red, and a
message displayed at the bottom of the field.
Task 3

In this task you will enable real-time preview of the user's comments
on the web page. This should be displayed in the same format as the
regular comments. You need to complete the following tasks:

Show a preview just above the form using the same format as the regular
comments on the page.
Display the preview only if the user has entered valid information into
the form, and the form is not in its pristine state.
Upon submitting the valid comment, the comment should join the regular
comments on the page. The date for the comment should be set
automatically upon submission of the form